+++ TLD-Twitch-Integration/Properties/AssemblyInfo.cs
@@ -3,7 +3,7 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("TLD-Twitch-Integration")]
 [assembly: AssemblyCopyright("B1gF1s4")]
-[assembly: AssemblyVersion("0.1.1")]
-[assembly: AssemblyFileVersion("0.1.1")]
-[assembly: MelonInfo(typeof(TLD_Twitch_Integration.Mod), "TLD-Twitch-Integration", "0.1.1", "B1gF1s4")]
+[assembly: AssemblyVersion("0.1.2")]
+[assembly: AssemblyFileVersion("0.1.2")]
+[assembly: MelonInfo(typeof(TLD_Twitch_Integration.Mod), "TLD-Twitch-Integration", "0.1.2", "B1gF1s4")]
 [assembly: MelonGame("Hinterland", "TheLongDark")]